Clean up, refactor
